+++ src/main/java/pl/bartekbak/vetclinicreservations/repository/VetRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.bartekbak.vetclinicreservations.entity.Vet;
 
 @Repository
-public interface VetRepository extends JpaRepository <Vet, Integer> {
+public interface VetRepository extends JpaRepository <Vet, Long> {
 }
